+++ PokerGameApp/PokerGameApp/Model/PokerGame.swift
@@ -87,81 +87,5 @@ class PokerGame {
     func judgeHighCard() -> [Int] {
         
         return [0,1,2,3,4]
-    }
-    
-}
-
-class Dealer {
-    private let cardNum: Int
-    private var round: Int
-    private var cardDeck: CardDeck
-    private(set) var myCard: [Card]
-    
-    init(cardNum: Int) {
-        self.cardNum = cardNum
-        self.round = 0
-        self.cardDeck = CardDeck()
-        cardDeck.shuffle()
-        self.myCard = []
-    }
-    // dealer should shuffle card (OK)
-    
-    func deal() -> [Card] {
-        switch (cardNum, round) {
-        case (5, 1):
-            return [dealOneCard(face: .down), dealOneCard(face: .up)]
-        case (7, 1):
-            return [dealOneCard(face: .down), dealOneCard(face: .down),  dealOneCard(face: .up)]
-        case (7, 5):
-            return [dealOneCard(face: .down)]
-        default:
-            return [dealOneCard(face: .up)]
-        }
-    }
-
-    
-    func dealOneCard(face: Face) -> Card {
-        var oneCard = cardDeck.takeOutCard()
-        oneCard.face = face
-        
-        return oneCard
-    }
-    
-    func getCard() {
-        round += 1
-        myCard += deal()
-    }
-    
-    func openCard() {
-        for i in 0..<myCard.count {
-            if myCard[i].face == .down {myCard[i].face = Face.up}
-        }
-    }
-}
-
-
-class Player {
-    private(set) var cards: [Card]
-    private(set) var betting: Int
-    let index: Int
-    
-    init(index: Int) {
-        self.cards = []
-        self.betting = 0
-        self.index = index
-    }
-    
-    func getCard(from cards: [Card]) {
-        self.cards += cards
-    }
-    
-    func bet(money: Int) {
-        betting += money
-    }
-    
-    func openCard() {
-        for i in 0..<cards.count {
-            if cards[i].face == .down {cards[i].face = Face.up}
-        }
-    }
+    }    
 }
